backend-main/requirements.txt
alembic
bcrypt
cryptography
Flask
Flask-Cors
Flask-SQLAlchemy
gunicorn
PyJWT
stripe
flask_migrate
psycopg2-binary
nltk
numpy
Pillow
requests
tensorflow
tika
pandas
python-docx
keras
spacy
spektral
pyception
chardet
python-dotenv